linked lists

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -13,7 +13,9 @@
 
 size_t print_listint(const listint_t *h)
 {
-	int count = 0;
+	size_t count = 0;
+
+
 	if (!h)
 	{
 		while (h)
